// src/shared/components/chart/echarts-for-react/models/interfaces/chart.interfaces.ts
export interface IChartData {
  // #region coordenadas cartesianas
  xAxis?: IxAxis;
  yAxis?: IyAxis;
  grid?: IGrid;
  // #endregion coordenadas cartesianas

  // #region coordenadas de grafico de area polar
  polar?: IPolar | IPolar[];
  radiusAxis?: IRadiusAxis | IRadiusAxis[];
  angleAxis?: IAngleAxis | IAngleAxis[];
  // #endregion coordenadas de grafico de area polar

  // #region grafico radar
  radar?: {
    indicator?: IIndicator[];
    shape?: TShape;
  };
  series?: ISeries | ISeries[];
  // #endregion grafico radar
}

// #region coordenadas cartesianas
interface IxAxis {
  type?: string;
  max?: number;
  splitLine?: ISplitLine;
  axisTick?: IAxisTick;
  axisLabel?: IStyleText;
  axisLine?: IAxisLine;
  data?: string[];
}
interface ISplitLine {
  show?: boolean;
}
interface IyAxis {
  type?: string;
  axisLabel?: IStyleText;
  data?: string[];
  axisTick?: IAxisTick;
  axisLine?: IAxisLine;
}
interface IAxisLine {
  show?: boolean;
}
interface IAxisTick {
  show?: boolean;
}
interface IGrid {
  left?: string;
  right?: string;
  top?: number;
  bottom?: number;
}
// #endregion coordenadas cartesianas

// -----------------------------------------------------------------------------------

// #region coordenadas de grafico de area polar
interface IPolar {
  radius?: (string | number)[] | number[] | string[];
}
export interface IRadiusAxis {
  max?: number;
  polarIndex?: number;
}
export interface IAngleAxis {
  type?: TType;
  startAngle?: number;
  data: string[];
  polarIndex?: number;
  endAngle?: number;
}
// #endregion coordenadas de grafico de area polar

// -----------------------------------------------------------------------------------

// #region grafico radar
export interface IIndicator {
  name?: string;
  max?: number;
}
export interface ISeries {
  name?: string;
  barWidth?: number;
  type?: ChartType;
  radius?: string | string[];
  center?: string[];
  startAngle?: number;
  endAngle?: number;
  avoidLabelOverlap?: boolean;
  label?: ILabel;
  emphasis?: {
    label?: ILabel;
  };
  labelLine?: ILabelLine;
  showBackground?: boolean;
  backgroundStyle?: IBackgroundStyle;
  coordinateSystem?: string;
  polarIndex?: number;

  data: (IDataChart | number)[] | IDataChart[] | number[];
  itemStyle?: IItemStyle;
}
export interface ILabel {
  show?: boolean;
  color?: string;
  fontSize?: number;
  fontWeight?: TFontWeight;
  position?: IPosition | number[];
  formatter?: string | ((...args: any[]) => string);
  align?: TAlign;
}
interface IBackgroundStyle {
  color?: string;
  borderRadius?: number | number[];
}
export interface ILabelLine {
  show?: boolean;
}
// #endregion grafico radar

// -----------------------------------------------------------------------------------

// #region informacion (data) de la grafica
export interface IDataChart {
  value: number | number[];
  name?: string;
  itemStyle?: IItemStyle;
}

export interface IItemStyle {
  color?: string;
  borderRadius?: number | number[];
  borderColor?: string;
  borderWidth?: number;
}
// ...
export interface IStyleText {
  show?: boolean;
  color?: string;
  fontWeight?: TFontWeight;
  fontSize?: number;
}